requirements.txt
django
djangorestframework
psycopg2
django-environ
pylint
flake8
pylint-django